Backend/requirements.txt
gunicorn==23.0.0
# Web Framework
APScheduler==3.11.2
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.21
sendgrid==6.12.5

# Database
psycopg2-binary==2.9.11
sqlalchemy==2.0.45
supabase==2.27.1

# Environment
python-dotenv==1.2.1

# CrewAI
crewai==1.6.1
crewai-tools==1.6.1

# LangChain
langchain==1.2.3
langchain-core==1.2.7
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langgraph==1.0.6
langchain-google-genai==4.1.3
langchain-litellm==0.3.5

# AI/LLM Providers
anthropic==0.75.0
google-generativeai==0.8.6
openai==2.15.0

# HTTP & API
google-search-results==2.4.2
requests==2.32.5
aiohttp==3.13.3
httpx==0.28.1

# Data Processing
pandas==2.3.3
numpy==2.4.1
PyMuPDF==1.26.7
scikit-learn==1.8.0

# LLM Gateway & Auth
litellm==1.80.15
email-validator==2.3.0
PyJWT==2.10.1
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
passlib==1.7.4
argon2-cffi==25.1.0
python-dateutil==2.9.0.post0
python-json-logger==4.0.0

# Task Management
jira==3.10.5

# Scheduler
apscheduler==3.11.2  

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
aiosqlite>=0.19.0
httpx>=0.27.0
pypdf==6.6.2
pytesseract==0.3.13
pdf2image==1.17.0